Trait nekoton_indexer_utils::ShardAccountsMapExt

pub trait ShardAccountsMapExt {
    fn find_account(&self, account: &UInt256) -> Result<Option<ExistingContract>>;
}

Helper trait to reduce boilerplate for getting accounts from shards state

Required Methods

fn find_account(&self, account: &UInt256) -> Result<Option<ExistingContract>>

Looks for a suitable shard and tries to extract information about the contract from it